% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetscoresmod-package.R
\docType{package}
\name{tweetscoresmod-package}
\alias{tweetscoresmod-package}
\alias{tweetscoresmod}
\title{Tools for the estimation of ideology scores with Twitter data}
\description{
This package provides a series of functions
to collect and analyze Twitter data, and to estimate the
in Barbera (2015, Political Analysis).
}
\seealso{
\code{\link{estimateIdeology}}, \code{\link{getFriends}},
}
\author{
Pablo Barbera \email{P.Barbera@lse.ac.uk}
}
